Revert to old scraping method for translations
The code using Google Translate's `batchexecute` API endpoint seem to be
giving us pretty poor (sometimes half-untranslated!) translations. We
revert to the old method for now. While this means we need two requests,
the APIs don't appear to be rate-limited, and it's better than the
current situation at least.

I also removed the commented-out `translate` method, because we can
always get it back through Git history, and it somewhat clutters up the
code.